Move more of the form into GameForm
